Move preprocessors into package
